Simplified configuration
