+++ src/Api/Utility/AuthConfig.cs
@@ -1,0 +1,9 @@
+﻿namespace PaderbornUniversity.SILab.Hip.CmsApi.Utility
+{
+    public class AuthConfig
+    {
+        public string Audience { get; set; }
+
+        public string Authority { get; set; }
+    }
+}
